requirements.txt
python-telegram-bot[webhooks]==21.1
gspread
oauth2client
openai
